% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jump.R
\docType{data}
\name{default_barrier_strings}
\alias{default_barrier_strings}
\title{Default barrier strings}
\format{
An object of class \code{character} of length 19.
}
\usage{
default_barrier_strings
}
\description{
Character vector of default barrier strings. Entries chosen based on default
RStudio autocreation of matching parenthesis and quotes.
}
\keyword{datasets}
