% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jump.R
\name{get-set-barrier-strings}
\alias{get-set-barrier-strings}
\alias{set_barrier_strings}
\alias{get_barrier_strings}
\title{Get and set barrier strings}
\usage{
set_barrier_strings(barrier_strings)

get_barrier_strings()
}
\arguments{
\item{barrier_strings}{A character vector for barrier strings (there should
be no empty strings) or \code{NULL} (to reset to default values).}
}
\value{
\code{set_barrier_strings()} returns call to \code{\link[=options]{options()}} after setting
barrier strings. \code{get_barrier_strings()} returns currently set barrier
strings (if nothing was set explicitly, it returns \link[=default_barrier_strings]{default barrier strings}).
}
\description{
These functions get and set 'cursorjumper' barrier strings which will be used
in \code{\link[=jump]{jump()}} for a decision making whether to perform a jump.
}
\details{
\strong{Note} that symbols in \code{barrier_strings} shouldn't be explicitly
escaped, i.e. they should be in the form present in RStudio.
}
\examples{
\dontrun{
# If nothing is set, it returns default barriers
barriers <- get_barrier_strings()

# Here " symbol shouldn't be explicitly escaped
set_barrier_strings(c("a", '"]'))
get_barrier_strings()

set_barrier_strings(barriers)
}
}
