% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jump.R
\name{jump}
\alias{jump}
\title{Jump over predefined barrier strings}
\usage{
jump()
}
\value{
Returns \code{TRUE} if there were no errors.
}
\description{
Jump over predefined barrier strings
}
\details{
Execute this function to perform jump in \link[rstudioapi:getActiveDocumentContext]{active document context}. It means that cursor
movement will be done in both source and console panels.

Jump is done if any \link[=set_barrier_strings]{set barrier strings} is found to
the right of the current cursor position (on the current line). If multiple
barrier strings are found then the first one is said "to be detected". After
that cursor is moved right after the detected barrier.

\strong{Note} that this function is usually helpful if binded to a keyboard
shortcut.
}
